consolidate into scene parser
